Remove more tags and former footer, header and banner
