+++ frontend/.env.sample
@@ -24,7 +24,6 @@ NAMED_MAPS_ENV=
 GFW_WIDGETS_BASE_URL=0.0.0.0:5000
 
 # feature flags
-DATA_FORM_ENABLED=true
 DATA_DOWNLOAD_ENABLED=true
 SHOW_WORLD_MAP_IN_EXPLORE=true
 ALWAYS_DISPLAY_DASHBOARD_INFO=true
@@ -33,6 +32,5 @@ DISABLE_PROFILES=false
 ENABLE_LEGACY_TOOL_SEARCH=false
 ENABLE_VERSIONING=false
 ENABLE_TOOL_PANEL=false
-CONSOLIDATE_INSIGHTS=true
 ENABLE_COUNTRY_PROFILES=true
 ENABLE_LOGISTIC_LAYERS_TAB=true
