+++ frontend/.eslintrc.json
@@ -65,7 +65,6 @@
     "ENABLE_INTERSECTION_OBSERVER": true,
     "ENABLE_VERSIONING": true,
     "ENABLE_TOOL_PANEL": true,
-    "CONSOLIDATE_INSIGHTS": true,
     "ENABLE_COUNTRY_PROFILES": true,
     "ENABLE_LOGISTIC_LAYERS_TAB": true,
     "API_V3_URL": true,
@@ -80,7 +79,6 @@
     "SHOW_WORLD_MAP_IN_EXPLORE": true,
     "ALWAYS_DISPLAY_DASHBOARD_INFO": true,
     "DATA_FORM_ENDPOINT": true,
-    "DATA_FORM_ENABLED": true,
     "PDF_DOWNLOAD_URL": true,
     "TRANSIFEX_API_KEY": true,
     "PERF_TEST": true,
