+++ frontend/docs/frontend/environments.md
@@ -69,7 +69,6 @@ REDUX_LOGGER_ENABLED=
 USE_SERVICE_WORKER=
 ALWAYS_DISPLAY_DASHBOARD_INFO=
 DATA_DOWNLOAD_ENABLED=
-DATA_FORM_ENABLED=
 ENABLE_DASHBOARDS=
 DISABLE_PROFILES=
 ENABLE_LEGACY_TOOL_SEARCH=
@@ -78,7 +77,6 @@ ENABLE_COOKIE_BANNER=
 ENABLE_INTERSECTION_OBSERVER=
 ENABLE_VERSIONING=
 ENABLE_TOOL_PANEL=
-CONSOLIDATE_INSIGHTS=
 ENABLE_COUNTRY_PROFILES=
 ENABLE_LOGISTIC_LAYERS_TAB=
 ```
