+++ frontend/docs/index.mdx
@@ -128,7 +128,6 @@ API_STORY_CONTENT: URL of the deep dive stories API
 API_SOCIAL: URL of the tweets API
 GOOGLE_ANALYTICS_KEY: API key for Google Analytics
 USER_REPORT_KEY: API key for User Report
-DATA_FORM_ENABLED: enable contact form in Data page
 DATA_FORM_ENDPOINT: end point to send form values in Data page
 PDF_DOWNLOAD_URL: end point to download a PDF snapshot of the page
 PERF_TEST: when set to true, enables performance evaluation mode. Only supported outside of production environments. See https://github.com/garbles/why-did-you-update for more info.
