+++ frontend/env.test.js
@@ -15,7 +15,6 @@ module.exports = {
   USE_SERVICE_WORKER: true,
   ALWAYS_DISPLAY_DASHBOARD_INFO: true,
   DATA_DOWNLOAD_ENABLED: true,
-  DATA_FORM_ENABLED: true,
   ENABLE_DASHBOARDS: true,
   DISABLE_PROFILES: false,
   ENABLE_LEGACY_TOOL_SEARCH: false,
@@ -23,7 +22,6 @@ module.exports = {
   ENABLE_INTERSECTION_OBSERVER: false,
   ENABLE_VERSIONING: true,
   ENABLE_TOOL_PANEL: true,
-  CONSOLIDATE_INSIGHTS: true,
   ENABLE_COUNTRY_PROFILES: true,
   ENABLE_LOGISTIC_LAYERS_TAB: false
 };
