+++ frontend/scripts/react-components/home/entrypoints/entrypoints.scss
@@ -2,7 +2,7 @@
 $entrypoint-slide-height: 550px;
 $entrypoints-stacked-card-size: 360px;
 
-.c-new-entrypoints {
+.c-entrypoints {
   height: 100%;
   padding-top: $nav-height;
