+++ frontend/scripts/react-components/nav/top-nav/cards/insights-card.component.jsx
