+++ frontend/scripts/react-components/nav/top-nav/tabs/tools-insights.component.jsx
