+++ frontend/scripts/react-components/nav/top-nav/top-nav.component.jsx
@@ -1,19 +1,25 @@
 import React, { Suspense } from 'react';
+import { Transition } from 'react-spring/renderprops';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 import throttle from 'lodash/throttle';
 import { NavLink } from 'redux-first-router-link';
 import NavLinks from 'react-components/nav/nav-links.component';
+import Heading from 'react-components/shared/heading';
 import LocaleSelector from 'react-components/nav/locale-selector/locale-selector.container';
 import Search from 'react-components/nav/global-search/global-search.container';
 import ToolSearch from 'react-components/tool/tool-search/tool-search.container';
 import Img from 'react-components/shared/img';
+import { BREAKPOINTS } from 'constants';
+import ResizeListener from 'react-components/shared/resize-listener.component';
 
 import 'scripts/react-components/nav/top-nav/top-nav.scss';
 
+import ToolsInsights from './tabs/tools-insights.component';
+
 const DownloadPdfLink = React.lazy(() => import('./download-pdf-link.component'));
 
-class TopNav extends React.PureComponent {
+class TopNavRedesign extends React.PureComponent {
   state = {
     backgroundVisible: false,
     menuOpen: false
@@ -21,14 +27,13 @@ class TopNav extends React.PureComponent {
 
   navLinkProps = {
     exact: false,
-    strict: false,
-    isActive: null
+    strict: false
   };
 
   mobileMenuRef = React.createRef(null);
 
   componentDidMount() {
-    window.addEventListener('click', this.handleClickOutside);
+    // window.addEventListener('click', this.handleClickOutside);
     window.addEventListener('scroll', this.setBackground, { passive: true });
   }
 
@@ -64,9 +69,52 @@ class TopNav extends React.PureComponent {
 
   handleToggleClick = () => this.setState(state => ({ menuOpen: !state.menuOpen }));
 
-  renderDesktopMenu() {
-    const { backgroundVisible } = this.state;
-    const { links, printable, showLogo, className, page } = this.props;
+  renderPopoverMenu() {
+    const { menuOpen } = this.state;
+    return (
+      <div className="nav-tabs-container">
+        <Transition
+          items={menuOpen}
+          from={{ opacity: 0, transform: 'translateY(-110%)' }}
+          leave={{ opacity: 0, transform: 'translateY(-110%)' }}
+          enter={{ opacity: 1, transform: 'translateY(0%)' }}
+        >
+          {show =>
+            show &&
+            (props => (
+              <div style={props}>
+                <ToolsInsights />
+              </div>
+            ))
+          }
+        </Transition>
+
+        <Transition
+          items={menuOpen}
+          from={{ opacity: 0 }}
+          leave={{ opacity: 0 }}
+          enter={{ opacity: 1 }}
+        >
+          {show =>
+            show &&
+            (props => (
+              <div
+                role="button"
+                tabIndex={0}
+                style={props}
+                onClick={() => this.setState({ menuOpen: false })}
+                className="-backdrop"
+              />
+            ))
+          }
+        </Transition>
+      </div>
+    );
+  }
+
+  renderMenu() {
+    const { links, printable, showLogo, page } = this.props;
+    const { menuOpen } = this.state;
     const allLinks = [];
 
     if (showLogo && !ENABLE_TOOL_PANEL) {
@@ -81,146 +129,132 @@ class TopNav extends React.PureComponent {
 
     allLinks.push(...links);
 
+    const SearchLabel = desktop => (desktop ? null : <span className="search-text">SEARCH</span>);
+
     return (
-      <div className="align-justify -desktop-menu">
-        {ENABLE_TOOL_PANEL && (
-          <div className="column medium-2">
-            <ul className="top-nav-item-list">
-              <li className="top-nav-item">
-                <NavLink
-                  exact
-                  strict
-                  to={{ type: 'home' }}
-                  className={cx('top-nav-link', '-new-logo')}
-                >
-                  <span className="top-nav-logo-container">
+      <ResizeListener>
+        {({ windowWidth }) => {
+          const isDesktop = windowWidth >= BREAKPOINTS.laptop;
+          const isSmall = windowWidth <= BREAKPOINTS.small;
+          return (
+            <div className="nav-menu">
+              <div className="first-row">
+                <div className="left-section">
+                  <button className="top-nav-toggle-btn" onClick={this.handleToggleClick}>
+                    <svg className={`icon icon-${menuOpen ? 'close' : 'menu'}`}>
+                      <use xlinkHref={`#icon-${menuOpen ? 'close' : 'menu'}`} />
+                    </svg>
+                  </button>
+                  <NavLink exact strict to={{ type: 'home' }} className="top-nav-logo">
                     <Img
+                      loading="lazy"
                       className="logo-image"
-                      src={
-                        className === '-light' || className === '-egg-shell' || backgroundVisible
-                          ? '/images/logos/new-logo-trase-red.svg'
-                          : '/images/logos/new-logo-trase-white.svg'
-                      }
-                      alt="trase"
+                      src="/images/logos/logo-trase-small.svg"
+                      alt="trase logo"
                     />
-                  </span>
-                </NavLink>
-              </li>
-            </ul>
-          </div>
-        )}
-        <div className="column medium-8">
-          <div className="top-nav-item-list-container">
-            <ul className="top-nav-item-list">
-              <NavLinks
-                links={allLinks}
-                itemClassName="top-nav-item"
-                linkClassName="top-nav-link"
-                linkActiveClassName="top-nav-link -active"
-                navLinkProps={this.navLinkProps}
-              />
-            </ul>
-          </div>
-        </div>
-        <div className="column medium-2">
-          <div className="top-nav-item-list-container -flex-end">
-            <ul className="top-nav-item-list">
-              <li className="top-nav-item">
-                <LocaleSelector />
-              </li>
-              <li className="top-nav-item">
-                {page === 'tool' ? <ToolSearch /> : <Search className="top-nav-search" />}
-              </li>
-              {printable && (
-                <li className="top-nav-item">
-                  <Suspense fallback={null}>
-                    <DownloadPdfLink />
-                  </Suspense>
-                </li>
+                    <Heading
+                      className="trase-logo-text"
+                      color="grey"
+                      as="span"
+                      variant="sans"
+                      size={isSmall ? 'rg' : 'lg'}
+                      weight="bold"
+                    >
+                      trase{' '}
+                    </Heading>
+                    <Heading
+                      color="pink"
+                      as="span"
+                      variant="sans"
+                      size={isSmall ? 'rg' : 'lg'}
+                      weight="bold"
+                    >
+                      supply chains
+                    </Heading>
+                  </NavLink>
+                </div>
+                <div className="right-section">
+                  {isDesktop && !menuOpen && (
+                    <ul className="nav-menu-main-navigation">
+                      <NavLinks
+                        links={allLinks}
+                        itemClassName="top-nav-item"
+                        linkClassName="top-nav-link"
+                        linkActiveClassName="top-nav-link -active"
+                        navLinkProps={this.navLinkProps}
+                      />
+                    </ul>
+                  )}
+                  <ul className="top-nav-item-list">
+                    <li className="top-nav-item search-container">
+                      <span className="search-icon">
+                        {page === 'tool' ? (
+                          <ToolSearch labelComponent={isDesktop ? SearchLabel : null} />
+                        ) : (
+                          <Search
+                            className="top-nav-search"
+                            labelComponent={isDesktop ? SearchLabel : null}
+                          />
+                        )}
+                      </span>
+                    </li>
+                    <li className="top-nav-item">
+                      <LocaleSelector />
+                    </li>
+                    {printable && (
+                      <li className="top-nav-item">
+                        <Suspense fallback={null}>
+                          <DownloadPdfLink />
+                        </Suspense>
+                      </li>
+                    )}
+                  </ul>
+                </div>
+              </div>
+              {!isDesktop && (
+                <div className="second-row">
+                  <ul className="nav-menu-main-navigation">
+                    <NavLinks
+                      links={allLinks}
+                      itemClassName="top-nav-item"
+                      linkClassName="top-nav-link"
+                      linkActiveClassName="top-nav-link -active"
+                      navLinkProps={this.navLinkProps}
+                    />
+                  </ul>
+                </div>
               )}
-            </ul>
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  renderMobileMenu() {
-    const { links, showLogo } = this.props;
-    const { menuOpen } = this.state;
-
-    const toggleBtnIcon = menuOpen ? 'close' : 'menu';
-
-    const allLinks = [];
-
-    allLinks.push(...links);
-
-    return (
-      <div className="row -mobile-menu" ref={this.mobileMenuRef}>
-        <div className="top-nav-bar column small-12">
-          <ul className="top-nav-item-list">
-            <li className="top-nav-item -no-margin">
-              <NavLink
-                exact
-                strict
-                to={{ type: 'home' }}
-                className={cx('top-nav-link', '-logo', { '-hide-when-on-top': !showLogo })}
-              >
-                <Img src="/images/logos/logo-trase-nav.png" alt="trase" />
-              </NavLink>
-            </li>
-            <li className="top-nav-item -no-margin">
-              <button className="top-nav-toggle-btn" onClick={this.handleToggleClick}>
-                <svg className={`icon icon-${toggleBtnIcon}`}>
-                  <use xlinkHref={`#icon-${toggleBtnIcon}`} />
-                </svg>
-              </button>
-            </li>
-          </ul>
-        </div>
-        {menuOpen && (
-          <div className="top-nav-collapse column small-12">
-            <ul className="top-nav-item-list-collapse">
-              <NavLinks
-                links={allLinks}
-                itemClassName="top-nav-item-collapse"
-                linkClassName="top-nav-link-collapse"
-                linkActiveClassName="top-nav-link-collapse -active"
-                navLinkProps={this.navLinkProps}
-              />
-              <li className="top-nav-item-collapse">
-                <LocaleSelector />
-              </li>
-            </ul>
-          </div>
-        )}
-      </div>
+            </div>
+          );
+        }}
+      </ResizeListener>
     );
   }
 
   render() {
-    const { className } = this.props;
+    const { className, page } = this.props;
     const { backgroundVisible, menuOpen } = this.state;
-
+    const YELLOW_PAGES = ['profiles', 'profile', 'data'];
     return (
       <div
         className={cx(
           'c-nav',
           {
             '-has-background': backgroundVisible || menuOpen,
+            '-yellow-background': YELLOW_PAGES.includes(page) && !backgroundVisible && !menuOpen,
             '-no-shadow': className === '-egg-shell'
           },
           className
         )}
       >
-        {this.renderDesktopMenu()}
-        {this.renderMobileMenu()}
+        {this.renderMenu()}
+        {this.renderPopoverMenu()}
       </div>
     );
   }
 }
 
-TopNav.propTypes = {
+TopNavRedesign.propTypes = {
   page: PropTypes.string,
   links: PropTypes.array,
   showLogo: PropTypes.bool,
@@ -229,8 +263,8 @@ TopNav.propTypes = {
   pageOffset: PropTypes.number
 };
 
-TopNav.defaultProps = {
+TopNavRedesign.defaultProps = {
   pageOffset: 0
 };
 
-export default TopNav;
+export default TopNavRedesign;
