+++ frontend/scripts/react-components/shared/cookie-banner/cookie-banner.scss
@@ -1,6 +1,6 @@
 @import 'styles/settings';
 
-.c-new-cookie-banner {
+.c-cookie-banner {
   position: fixed;
   left: 0;
   right: 0;
