+++ frontend/scripts/react-components/shared/footer/footer.scss
@@ -1,6 +1,6 @@
 @import 'styles/settings';
 
-.c-new-footer {
+.c-footer {
   position: relative;
   z-index: $z-above;
   padding: 48px 0 56px;
